fix(llm-agents): guard Google LM test against model-selection drop race, restore @stable (#596)
Root cause: the Language Model node's default resolves through a missing
__default_language_model__ variable and falls back to the first configured
provider's model; a custom_component/update racing the setupGoogle selection
reverts the node to that fallback (the #491 class), so the build runs an
unintended model and 'built successfully' never fires. Fix is test-side per
the issue decision: pin a deterministic Gemini flash model (resolveGeminiModel
extracted to a shared helper) and add a bounded pre-run widget gate that
re-applies a dropped selection. Asserts and the 30s toast timeout untouched;
@stable restored on the test.
